cpu: Clean up immediate address handling
After further inspection, none of the opcodes that have an immediate
addressing option ever write to the address they are passed. Due to
this, the operand set by the immediate addressing function can be
changed to the currenc PC, allowing it to be accessed by instruction
functions the same way as non-immediate addressing.